chore(ref): Modularize and reformat code
