+++ lib/home.dart
@@ -8,6 +8,8 @@ import 'package:flutter_tts/flutter_tts.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:helping_hands/currency_detection/currency.dart';
 import 'package:helping_hands/image_captioning/image_captioning.dart';
+import 'package:helping_hands/ocr/dialog_ocr.dart';
+import 'package:helping_hands/sos/sos_dialog.dart';
 import 'package:hexcolor/hexcolor.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:shake/shake.dart';
@@ -38,10 +40,6 @@ class _HomePageState extends State<HomePage> {
   File _capImage;
   File _currImage;
   final FlutterTts flutterTts = FlutterTts();
-  TextEditingController _controller1 = new TextEditingController();
-  TextEditingController _controller2 = new TextEditingController();
-  TextEditingController _controller3 = new TextEditingController();
-  TextEditingController _controller4 = new TextEditingController();
   final Telephony telephony = Telephony.instance;
 
   PageController _controller = PageController(
@@ -151,6 +149,8 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     Size screen = MediaQuery.of(context).size;
+    sosDialog sd = new sosDialog();
+    ocrDialog od = new ocrDialog();
     return Scaffold(
       body: PageView(
         controller: _controller,
@@ -193,7 +193,7 @@ class _HomePageState extends State<HomePage> {
                       child: FlatButton(
                           highlightColor: Hexcolor('#A8DEE0'),
                           splashColor: Hexcolor('#F9E2AE'),
-                          onPressed: () => _optionsDialogBox(),
+                          onPressed: () => od.optionsDialogBox(context),
                           child: Text("Text Extraction from Images",
                               style: TextStyle(
                                   fontSize: 27.0,
@@ -219,7 +219,7 @@ class _HomePageState extends State<HomePage> {
                       child: FlatButton(
                           highlightColor: Colors.yellow[900],
                           splashColor: Colors.yellow[500],
-                          onPressed: () => _sosDialogBox(),
+                          onPressed: () => sd.sosDialogBox(context),
                           child: Text("SOS Settings",
                               style: TextStyle(
                                   fontSize: 27.0,
@@ -232,252 +232,6 @@ class _HomePageState extends State<HomePage> {
         physics: BouncingScrollPhysics(),
       ),
     );
-  }
-
-  Future _speak() async {
-    await flutterTts.speak("something");
-  }
-
-  Future<void> _sosDialogBox() {
-    return showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-              title: Text("Enter phone numbers you would like to contact"),
-              content: new SingleChildScrollView(
-                  child: new ListBody(children: <Widget>[
-                TextFormField(
-                  controller: _controller4,
-                  decoration: InputDecoration(
-                    labelText: 'Enter your name:',
-                  ),
-                ),
-                TextFormField(
-                  keyboardType: TextInputType.number,
-                  controller: _controller1,
-                  decoration: InputDecoration(
-                    labelText: 'Enter phone number 1:',
-                  ),
-                ),
-                TextFormField(
-                  keyboardType: TextInputType.number,
-                  controller: _controller2,
-                  decoration: InputDecoration(
-                    labelText: 'Enter phone number 2:',
-                  ),
-                ),
-                TextFormField(
-                  keyboardType: TextInputType.number,
-                  controller: _controller3,
-                  decoration: InputDecoration(
-                    labelText: 'Enter phone number 3:',
-                  ),
-                ),
-                new SizedBox(height: 10),
-                new RaisedButton(
-                  onPressed: () {
-                    setNumbers((_controller1.text), (_controller2.text),
-                        (_controller3.text), _controller4.text);
-                  },
-                  color: Hexcolor('eaac8b'),
-                  child: Text(
-                    "Save Information",
-                    style: TextStyle(
-                        color: Colors.white, fontWeight: FontWeight.bold),
-                  ),
-                  elevation: 5.0,
-                  shape: new RoundedRectangleBorder(
-                      borderRadius: new BorderRadius.circular(16.0)),
-                )
-              ])));
-        });
-  }
-
-  setNumbers(String n1, String n2, String n3, String n) async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    print(n1);
-    print(n2);
-    print(n3);
-    await prefs.setString('n1', n1);
-    await prefs.setString('n2', n2);
-    await prefs.setString('n3', n3);
-    await prefs.setString('name', n);
-  }
-
-  Future<void> _optionsDialogBox() {
-    return showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            content: new SingleChildScrollView(
-              child: new ListBody(
-                children: <Widget>[
-                  RichText(
-                    text: TextSpan(
-                      style: Theme.of(context).textTheme.bodyText2,
-                      children: [
-                        WidgetSpan(
-                          child: Padding(
-                            padding:
-                                const EdgeInsets.symmetric(horizontal: 3.0),
-                            child: Icon(FlutterIcons.photo_camera_mdi),
-                          ),
-                        ),
-                        TextSpan(
-                          text: 'Choose mode',
-                          style: new TextStyle(
-                              fontWeight: FontWeight.bold, fontSize: 20),
-                        ),
-                      ],
-                    ),
-                  ),
-                  Padding(
-                    padding: EdgeInsets.all(8.0),
-                  ),
-                  GestureDetector(
-                    child: new Text('Take a picture'),
-                    onTap: openCamera,
-                  ),
-                  Padding(
-                    padding: EdgeInsets.all(8.0),
-                  ),
-                  GestureDetector(
-                    child: new Text('Select from gallery'),
-                    onTap: openGallery,
-                  ),
-                ],
-              ),
-            ),
-          );
-        });
-  }
-
-  Future<void> openCamera() async {
-    ImagePicker ip = new ImagePicker();
-    var picture = await ip.getImage(
-      source: ImageSource.camera,
-    );
-    var _extractText = await SimpleOcrPlugin.performOCR(picture.path);
-    print(_extractText.substring(20));
-    _speakOCR(_extractText.substring(20, _extractText.length - 15));
-    showOCRDialog(
-        _extractText.substring(20, _extractText.length - 15), picture);
-  }
-
-  Future<void> openGallery() async {
-    ImagePicker ip = new ImagePicker();
-    var picture = await ip.getImage(
-      source: ImageSource.gallery,
-    );
-    var _extractText = await SimpleOcrPlugin.performOCR(picture.path);
-    print(_extractText.substring(20));
-    _speakOCR(_extractText.substring(20, _extractText.length - 15));
-    showOCRDialog(
-        _extractText.substring(20, _extractText.length - 15), picture);
-  }
-
-  Future<void> showOCRDialog(String text, PickedFile picture) async {
-    final pngByteData = await picture.readAsBytes();
-    showGeneralDialog(
-        barrierColor: Colors.black.withOpacity(0.5),
-        transitionBuilder: (context, a1, a2, widget) {
-          return Transform.scale(
-            scale: a1.value,
-            child: Opacity(
-              opacity: a1.value,
-              child: AlertDialog(
-                shape: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(16.0)),
-                title: Text('Text Detected'),
-                content: SingleChildScrollView(
-                  child: Column(
-                    children: <Widget>[
-                      new Container(
-                        width: 300.0,
-                        height: 150,
-                        child: RaisedButton(
-                          onPressed: _stopTts,
-                          padding: const EdgeInsets.all(10.0),
-                          child: const Text('Stop',
-                              style: TextStyle(
-                                  color: Colors.white,
-                                  fontWeight: FontWeight.bold)),
-                          color: Hexcolor('b56576'),
-                          elevation: 5.0,
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(16.0)),
-                        ),
-                      ),
-                      new Container(height: 10),
-                      new Container(
-                        width: 300.0,
-                        height: 150,
-                        child: RaisedButton(
-                          onPressed: () {
-                            _speakOCR(text);
-                          },
-                          padding: const EdgeInsets.all(10.0),
-                          child: const Text('Replay',
-                              style: TextStyle(
-                                  color: Colors.white,
-                                  fontWeight: FontWeight.bold)),
-                          color: Hexcolor('b56576'),
-                          elevation: 5.0,
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(16.0)),
-                        ),
-                      ),
-                      new Container(height: 10),
-                      new Container(
-                        width: 300.0,
-                        height: 150,
-                        child: RaisedButton(
-                          onPressed: _pauseTts,
-                          padding: const EdgeInsets.all(10.0),
-                          child: const Text('Pause',
-                              style: TextStyle(
-                                  color: Colors.white,
-                                  fontWeight: FontWeight.bold)),
-                          color: Hexcolor('b56576'),
-                          elevation: 5.0,
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(16.0)),
-                        ),
-                      ),
-                      new Container(height: 10),
-                      new Image.memory(pngByteData),
-                      SizedBox(width: 20),
-                      new Text("$text"),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          );
-        },
-        transitionDuration: Duration(milliseconds: 200),
-        barrierDismissible: true,
-        barrierLabel: '',
-        context: context,
-        pageBuilder: (context, animation1, animation2) {});
-  }
-
-  Future _stopTts() async {
-    if (await Vibration.hasVibrator()) {
-      Vibration.vibrate(amplitude: 100, duration: 200);
-    }
-    flutterTts.stop();
-  }
-
-  void _pauseTts() {
-    flutterTts.pause();
-  }
-
-  Future _speakOCR(String text) async {
-    if (await Vibration.hasVibrator()) {
-      Vibration.vibrate(amplitude: 500, duration: 200);
-    }
-    await flutterTts.speak(text);
   }
 
   _speakPage(int a) async {
@@ -506,4 +260,3 @@ class _HomePageState extends State<HomePage> {
     }
   }
 }
-//
